requirements.txt
Django==1.8
python-dateutil==2.4.2
pytz==2015.2
six==1.9.0
wsgiref==0.1.2